+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Remove Pamac workaround since we have fixed the 'libpamac-full' AUR package upstream.
 - Add ReplaySorcery for use with GOverlay.
 - Add vkBasalt for use with GOverlay.
+- Change the 'linux-neptune' package to be 'linux-steamos'.
 
 ## 3.1.1 - The Boot Fix Update